Added auth to the template

+++ common/src/utils.rs
@@ -5,6 +5,6 @@ pub fn is_valid_email(email: &str) -> bool {
     email.contains('@') && email.contains('.')
 }
 
-pub fn find_contact_by_id(contacts: &[ContactDto], id: u32) -> Option<&ContactDto> {
+pub fn find_contact_by_id(contacts: &[ContactDto], id: i64) -> Option<&ContactDto> {
     contacts.iter().find(|contact| contact.id == Some(id))
 }
